refactor: make the library idiomatic and gate it with golangci-lint (#12)
* refactor: make the library idiomatic and gate it with golangci-lint

Sweeps every module for non-idiomatic Go and removes all inline lint
suppressions, so the standard is enforced in one place instead of waived
at 65 call sites. The complexity waivers are retired by actually
splitting the SVG path, CSS rule and CSS selector parsers into named
steps, and every integer conversion gosec flagged is now provably in
range rather than annotated.

Fixes found on the way: Pdf.Save created files with os.ModePerm (0777),
AcroForm.Add returned a nil receiver instead of nil, and internal/svg
tested for NaN with `x != x`.


* fix(ci): use golangci-lint-action v7 for golangci-lint v2

Action v6 rejects v2 version strings, so the lint job failed on its
first run.


---------

+++ internal/svg/budget_test.go
@@ -108,5 +108,5 @@ func TestRasterize_ClampsOversizedFontSize(t *testing.T) {
 	// An unclamped face rasterises a mask of its own em size: this input used to
 	// allocate ~2.7GB for a single glyph on a 118x118 canvas.
 	allocatedMB := (after.TotalAlloc - before.TotalAlloc) / (1 << 20)
-	assert.Less(t, int(allocatedMB), 64)
+	assert.Less(t, allocatedMB, uint64(64))
 }
